+++ src/GPAcalculator.java
@@ -2,13 +2,14 @@ public class GPAcalculator implements GPAcalc{
     /*TODO:
      * 1)Set up Test Cases for the calc(OPTIONAL)
      * 2)Implement the GPAcalc
-     *
      */
+
+
     @Override
-    public float GPAcalc(int[] data) {
+    public int GPAcalc(float[] data) {
         /**
-         * @param- int[] - [numOfcourse,numHDs,numDs,numCRs,numPs]
-         * */
+         * @param - float[] - [Estd GPA,numOfcourse,numHDs,numDs,numCRs,numPs]
+         * @returns - The total number of points required to achieve to Estd GPA */
         return 0;
     }
